# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "realt backend api",
    "description": "realt backend",
    "version": "1.0.0"
  },
  "host": "realt-backend.herokuapp.com",
  "basePath": "",
...
  "schemes": ["https"],
  "paths": {
...
    "/owners": {
      "post": {
        "tags": ["owners"],
        "summary": "create new owner",
        "parameters": [
          {
            "in": "body",
            "name": "owner",
            "schema": {
              "$ref": "#/definitions/Owner"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": ["owners"],
        "summary": "get owner",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": ["owners"],
        "summary": "update owner",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "owner",
            "schema": {
              "$ref": "#/definitions/OwnerUpdate"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admins": {
      "post": {
        "tags": ["admins"],
        "summary": "create new admin",
        "parameters": [
          {
            "in": "body",
            "name": "admin",
            "schema": {
              "$ref": "#/definitions/AdminWithKey"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": ["admins"],
        "summary": "get admin",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": ["admins"],
        "summary": "update admin",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "admin",
            "schema": {
              "$ref": "#/definitions/AdminUpdate"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/owners/auction": {
      "post": {
        "tags": ["auction"],
        "summary": "create new auction",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "auction",
            "schema": {
              "$ref": "#/definitions/Auction"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": ["auction"],
        "summary": "get auction",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": ["auction"],
        "summary": "update auction",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "auction",
            "schema": {
              "$ref": "#/definitions/Auction"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": ["auction"],
        "summary": "delete auction",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "auction",
            "schema": {
              "$ref": "#/definitions/Auction Id"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["auction"],
        "summary": "start auction",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "auction",
            "schema": {
              "$ref": "#/definitions/Auction Id"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "auth user, get token",
        "parameters": [
          {
            "in": "body",
            "name": "auth",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "refresh access token, get access token",
        "parameters": [
          {
            "in": "body",
            "name": "auth",
            "schema": {
              "$ref": "#/definitions/Refresh"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Owner": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordRepeat": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password",
        "passwordRepeat",
        "name",
        "surname",
        "phone"
      ]
    },
    "OwnerUpdate": {
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "required": ["name", "surname", "phone"]
    },
    "AdminWithKey": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordRepeat": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "keyDataCandidate": {
          "type": "object",
          "properties": {
            "key": {
              "type": "string"
            },
            "id": {
              "type": "string"
            }
          }
        }
      },
      "required": ["email", "password", "passwordRepeat", "name", "surname", "phone", "key"]
    },
    "AdminUpdate": {
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      },
      "required": ["name", "surname", "phone"]
    },
...
    "Auction": {
      "properties": {
        "address": {
          "type": "object",
          "properties": {
            "region": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "district": {
              "type": "string"
            },
            "street": {
              "type": "string"
            },
            "house": {
              "type": "string"
            },
            "apartment": {
              "type": "string"
            }
          }
        },
        "technicalSpec": {
          "type": "object",
          "properties": {
            "area": {
              "type": "number"
            },
            "balcony": {
              "type": "number"
            },
            "isBalconyIncluded": {
              "type": "boolean"
            },
            "rooms": {
              "type": "number"
            },
            "kitchenArea": {
              "type": "number"
            },
            "combinedBathroom": {
              "type": "boolean"
            },
            "separateRooms": {
              "type": "boolean"
            },
            "houseBuildYear": {
              "type": "number"
            },
            "houseBuildingMaterial": {
              "type": "string"
            },
            "floor": {
              "type": "string"
            },
            "conditionOfTheApartment": {
              "type": "string"
            },
            "landscapedHouse": {
              "type": "boolean"
            }
          }
        },
        "legalIssues": {
          "type": "object",
          "properties": {
            "yearsOfOwnership": {
              "type": "number"
            },
            "owners": {
              "type": "number"
            },
            "minorOwner": {
              "type": "boolean"
            },
            "basisOfOwnership": {
              "type": "string"
            },
            "encumbrances": {
              "type": "string"
            },
            "debts": {
              "type": "string"
            },
            "homePurpose": {
              "type": "string"
            }
          }
        },
        "auctionId": {
          "type": "string"
        }
      },
      "required": ["address", "technicalSpec", "legalIssues"]
    },
...
    "Login": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["email", "password"]
    },
    "Refresh": {
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": ["token"]
    }
  }
}